Enable -fms-extensions, fix LONG type, forward declarations
- Add -fms-extensions to Clang on macOS for MSVC compatibility
  (friend class as forward declaration, anonymous structs, etc.)
- Revert LONG to int32_t (matching Windows ABI) to fix VESSEL API
  method signature mismatches on ARM64
- Fix Panel.h/cpp Point2Screen to use LONG instead of long
- Add forward declarations for inner classes in AerodynSubsys.h
  (AerodynSelectorDial, AirbrakeLever, ElevatorTrimWheel)

Module build status: 18 .dylib modules link successfully:
- Celestial bodies: Vsop87, Moon, EarthAtm2006, MarsAtm2006, VenusAtm2006
- Vessels: HST, Quadcopter, ShuttlePB, ShuttleA_PL, Solarsail,
  ScriptVessel, MMU
- Plugins: FlightData, Framerate, Rcontrol, Notes, ExtMFD, MFDTemplate

42 module .cpp files still have Clang-specific compilation issues
(implicit type conversions, friend class scoping) that need
incremental fixes.

+++ Orbitersdk/include/OrbiterPlatform.h
@@ -39,7 +39,7 @@
 // Basic Windows types
 // -----------------------------------------------------------
 typedef uint32_t       DWORD;
-typedef long           LONG;
+typedef int32_t        LONG;
 typedef int64_t        LONGLONG;
 typedef uint16_t       WORD;
 typedef unsigned char  BYTE;
